% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_package.R
\name{fit.NB.st}
\alias{fit.NB.st}
\title{Fit space-time additive negative-binomial regression to 
estimate unrecognized deaths attributable to certain infectious disease
based on spatial-temporal data}
\usage{
fit.NB.st(
  formula,
  data,
  offset = NULL,
  niter = 5000,
  burn_in = 2500,
  rand.int = TRUE,
  str.int = NULL,
  slope.tvarying = FALSE,
  str.tvarying = NULL,
  covariate.AD,
  countsonly = FALSE,
  spatial.str,
  adj.mat.s = NULL,
  ID.spacetime = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" specifying the model to be fitted
using a symbolic description.}

\item{data}{a data frame containing variables used for fitting models.}

\item{offset}{an optional argument specifying an \emph{a prior} known component to be
included in the linear predictor. For example, log of at-risk populations.}

\item{niter}{a number specifying the number of MCMC iterations, the default is 5000.}

\item{burn_in}{a number specifying the number of iterations that is discarded, the default is 2500.}

\item{rand.int}{a logical value indicating whether time-specific random effects are included.}

\item{str.int}{a list containing two character strings when \code{rand.int} = \code{TRUE}.
The first character string specifies the variable indicating the clusters of
time-specific random effects, this variable must be included in the \code{data};
the second character string specifies the structure of time-specific random effects,
\code{AR1} and \code{AR2} fit random walk model of order 1 or 2,
and \code{exch} specifies exchangeable time-specific random effects.}

\item{slope.tvarying}{a logical value indicating whether time-varying coefficients are included.}

\item{str.tvarying}{a list containing vectors including three character strings
when \code{slope.tvarying} = \code{TRUE}.
The first character string specifies which variable having time-varying regression coefficient;
the second character string specifies the variable indicating the clusters of
time-varying coefficient, this variable must be included in the \code{data};
the third character string specifies the structure of time-varying coefficient,
\code{AR1} and \code{AR2} fit random walk model of order 1 or 2,
and \code{exch} specifies exchangeable random effects for time-varying coefficients.}

\item{covariate.AD}{a vector containing character strings specifying which variables need
to be set to 0 to estimate attributable deaths.}

\item{countsonly}{a logical value indicating whether posterior samples of random effects are
returned when \code{rand.int} = \code{TRUE} and/or \code{slope.tvarying} = \code{TRUE}, 
the default is \code{FALSE}.}

\item{spatial.str}{a list containing two character strings. The 
first character string specifies the variable indicating the clusters of
location-specific random effects, this variable must be included in the \code{data};
the second character string specifies the structure of location-specific random effects,
\code{exch} specifies exchangeable location-specific random effects, 
and \code{CAR} fits conditional autoregressive model (CAR).}

\item{adj.mat.s}{an optional matrix, when the CAR model is fitted, 
an between-location adjacency matrix must to be provided.}
}
\value{
This function returns a list containing:\tabular{ll}{
   \code{beta} \tab a matrix containing posterior samples of fixed effects and mean of time-varying 
   coefficients when \code{slope.tvarying} = \code{TRUE}. \cr
   \tab \cr
   \code{parm} \tab a matrix containing posterior samples of the over-dispersion parameter \code{xi}, 
   parameters controlling temporal dependency of time-specific random effects 
   when \code{rand.int} = \code{TRUE}, and parameters controlling spatial dependency of 
   location-specific random effects. When \code{exch} is specified for time-specific random effects, 
   \code{tau2.t} is included; when \code{AR1} or \code{AR2} are specified for time-specific random effects, 
   \code{tau2.t} and \code{rho.t} are included. When \code{exch} is specified for 
   location-specific random effects, \code{tau2.s} is included; 
   when \code{CAR} is specified for location-specific random effects, 
   \code{tau2.s} and \code{rho.s} are incldued.\cr
   \tab \cr
   \code{pred.counts} \tab a matrix containing posterior samples with burn-in samples are discarded 
   of observed counts. \cr
   \tab \cr
   \code{pred.null.counts} \tab a matrix containing posterior samples with burn-in samples are discarded 
   of estimated counts assuming variables specified using \code{covariate.AD} equal to 0. \cr
   \tab \cr
   \code{AD} \tab a matrix containing posterior samples with burn-in samples are discarded 
   of attributable deaths 
   (i.e., point-wise difference between \code{pred.counts} and \code{pred.null.counts}). \cr
   \tab \cr
   \code{accept.rate} \tab acceptance rate for updating the over-dispersion parameter 
   in Metropolis-Hasting algorithm. \cr
   \tab \cr
   \code{WAIC} \tab the widely applicable information criterion (WAIC) 
   in deviance scale that measures how the model fit to the observed data. \cr
   \tab \cr
   \code{lppd} \tab the log point-wise predictive probability of the 
   observed data under the fitted model. \cr
   \tab \cr
   \code{pWAIC} \tab the approximated effective number of parameters; \cr
}
when \code{countsonly} = \code{FALSE}, this function also returns the following: \tabular{ll}{
   \code{rand.int.t} \tab a matrix containing posterior samples of time-specific random effects
    when \code{rand.int} = \code{TRUE}. \cr
   \tab \cr
   \code{rand.int.s} \tab a matrix containing posterior samples of 
   location-specific random effects. \cr
   \tab \cr
   \code{beta.tvarying} \tab a list containing matrices of posterior samples 
   of time-varying coefficients when \code{slope.tvarying} = \code{TRUE}. \cr
   \tab \cr
   \code{parm.beta.tvarying} \tab a list containing matrices of posterior samples of 
   parameters controlling temporal dependency of time-varying coefficients when 
   \code{slope.tvarying} = \code{TRUE}; when \code{exch} is specified, \code{sigma2.X} is included; 
   when \code{AR1} or \code{AR2} are specified, \code{sigma2.X} and \code{rho.X} are included, 
   where \code{X} is replaced by variable names specified via \code{covariate.AD}. \cr
}
}
\description{
This function fits space-time additive negative binomial regression models 
using measures of virus transmission
as covariates to estimate unrecognized deaths attributable to infections caused 
by those virus
in epidemiological studies collecting data with spatial-temporal structures.
}
\examples{
data(toydata)

# Assume time-specific random effects and 
# assume time-varying coefficients
fit.st <- fit.NB.st(formula = y ~ lag1 + lag2,
data = dat.fit.NB,
offset = logPop,
niter = 5000, burn_in = 2500,
rand.int = TRUE,
str.int = list("week", "exch"),
slope.tvarying = TRUE, 
str.tvarying = list(c("lag1", "week", "exch"),
                    c("lag2", "week", "exch")),
covariate.AD = c("lag1", "lag2"),
spatial.str = list("states", "exch"), adj.mat.s = NULL)

}
\references{
Gelman, A., Hwang, J., & Vehtari, A. (2014). Understanding predictive 
information criteria for Bayesian models. 
\emph{Statistics and computing, 24}(6), 997-1016.
}
