% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_package.R
\name{fit.NB.timeseries}
\alias{fit.NB.timeseries}
\title{Fit negative-binomial regression to estimate unrecognized deaths attributable to
certain infectious disease}
\usage{
fit.NB.timeseries(
  formula,
  data,
  offset = NULL,
  niter = 5000,
  burn_in = 2500,
  rand.int = TRUE,
  str.int = NULL,
  covariate.AD,
  countsonly = FALSE,
  ID.spacetime = NULL,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" specifying the model to be fitted
using a symbolic description.}

\item{data}{a data frame containing variables used for fitting models.}

\item{offset}{an optional argument specifying an \emph{a prior} known component to be
included in the linear predictor. For example, log of at-risk populations.}

\item{niter}{a number specifying the number of MCMC iterations, the default is 5000.}

\item{burn_in}{a number specifying the number of iterations that is discarded, the default is 2500.}

\item{rand.int}{a logical value indicating whether time-specific random effects are included.}

\item{str.int}{a list containing two character strings when \code{rand.int} = \code{TRUE}.
The first character string specifies the variable indicating the clusters of
time-specific random effects, this variable must be included in the \code{data};
the second character string specifies the structure of time-specific random effects,
\code{AR1} and \code{AR2} fit random walk model of order 1 or 2,
and \code{exch} specifies exchangeable time-specific random effects.}

\item{covariate.AD}{a vector containing character strings specifying which variables need
to be set to 0 to estimate attributable deaths.}

\item{countsonly}{a logical value indicating whether posterior samples of random effects are
returned when \code{rand.int} = \code{TRUE}, the default is \code{FALSE}.}

\item{ID.spacetime}{an optional vector containing variable names specifying location and time indicators of 
observations, those indicators have to be included in the \code{data}, the default is \code{NULL}. 
Specifying location and time indicators is important if the user want to obtain estimates of 
attributable deaths across a certain time period and locations.}
}
\value{
This function returns a list containing:\tabular{ll}{
   \code{beta} \tab a matrix containing posterior samples of fixed effects and mean of time-varying 
   coefficients when \code{slope.tvarying} = \code{TRUE}. \cr
   \tab \cr
   \code{parm} \tab a matrix containing posterior samples of the over-dispersion parameter \code{xi}, 
   and parameters controlling temporal dependency of time-specific random effects when \code{rand.int} = \code{TRUE}. 
   When \code{exch} is specified for time-specific random effects, \code{tau2.t} is included; 
   when \code{AR1} or \code{AR2} are specified, \code{tau2.t} and \code{rho.t} are included. \cr
   \tab \cr
   \code{pred.counts} \tab a matrix containing posterior samples with burn-in samples are discarded 
   of observed counts. \cr
   \tab \cr
   \code{pred.null.counts} \tab a matrix containing posterior samples with burn-in samples are discarded 
   of estimated counts assuming variables specified using \code{covariate.AD} equal to 0. \cr
   \tab \cr
   \code{AD} \tab a matrix containing posterior samples with burn-in samples are discarded 
   of attributable deaths 
   (i.e., point-wise difference between \code{pred.counts} and \code{pred.null.counts}). \cr
   \tab \cr
   \code{accept.rate} \tab acceptance rate for updating the over-dispersion parameter 
   in Metropolis-Hasting algorithm. \cr
   \tab \cr
   \code{WAIC} \tab the widely applicable information criterion (WAIC) 
   in deviance scale that measures how the model fit to the observed data. \cr
   \tab \cr
   \code{lppd} \tab the log point-wise predictive probability of the 
   observed data under the fitted model. \cr
   \tab \cr
   \code{pWAIC} \tab the approximated effective number of parameters; \cr
}
when \code{countsonly} = \code{FALSE}, this function also returns the following: \tabular{ll}{
   \code{rand.int.t} \tab a matrix containing posterior samples of time-specific random effects
    when \code{rand.int} = \code{TRUE}. \cr
   \tab \cr
}
}
\description{
This function fits negative binomial regression models using measures of virus transmission
as covariates to estimate unrecognized deaths attributable to infections caused by those virus
in epidemiological studies with a time-series design.
}
\examples{
data(toydata)

# NO time-specific random effects 
fit.1 <- fit.NB.timeseries(formula = y ~ lag1 + lag2, 
data = toydata, offset = logPop, 
niter = 5000, burn_in = 2500,
rand.int = FALSE,
covariate.AD = c("lag1", "lag2"),
countsonly = TRUE) 

# Assume time-specific random effects modeled 
# via random walk model of order 1
fit.2 <- fit.NB.timeseries(formula = y ~ lag1 + lag2, 
data = toydata, offset = logPop, 
niter = 5000, burn_in = 2500,
rand.int = TRUE, str.int = list("week", "AR1"),
covariate.AD = c("lag1", "lag2"),
countsonly = TRUE) 

}
\references{
Gelman, A., Hwang, J., & Vehtari, A. (2014). Understanding predictive 
information criteria for Bayesian models. 
\emph{Statistics and computing, 24}(6), 997-1016.
}
