\name{NBRegAD-package}
\alias{NBRegAD-package}
\alias{NBRegAD}
\docType{package}
\title{Estimate unrecognized deaths attributable to certain infectious disease}
\description{
This package implement negative-binomial regression models to estimate unrecognized deaths attributable to
certain infectious disease based on data with temporal structure or spatial-temporal structure.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NBRegAD}
\packageIndices{NBRegAD}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Yuzi Zhang

Maintainer: Yuzi Zhang <yuzi.zhang@emory.edu>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
