% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_package.R
\name{sumcounts}
\alias{sumcounts}
\title{Summarize results from estimating attributable deaths}
\usage{
sumcounts(counts.post, ID.time, ID.loc, ID.aggre, timeperiod, locs)
}
\arguments{
\item{counts.post}{a matrix of posterior samples of counts, this matrix also
has to include location and time index of each counts in a wide format.}

\item{ID.time}{a character string specifying the time index of posterior samples.}

\item{ID.loc}{a character string specifying the location index of posterior samples.}

\item{ID.aggre}{a character string specifying the grouping variable.}

\item{timeperiod}{a vector specifying the time period of interest.}

\item{locs}{a vector specifying locations of interest.}
}
\value{
This function returns a data frame containing:\tabular{ll}{
   \code{est} \tab posterior mean of estimated counts. \cr
   \tab \cr
   \code{sd} \tab posterior standard deviation of estimated counts.\cr
   \tab \cr
   \code{lci} \tab lower bound of 95\% credible interval for estimated counts. \cr
   \tab \cr
   \code{uci} \tab upper bound of 95\% credible interval for estimated counts. \cr
}
}
\description{
This function summarizes results of negative-binomial models fitted to 
the observed data to obtain posterior mean, posterior standard deviation, and 
95\% credible interval of estimated counts  across time-period of interest 
and at locations of interest.
}
\examples{
data(toydata)

### Note that this example is just for the illustration, 
### the model doesn't fit the data well
fit <- fit.NB.timeseries(formula = y ~ lag1 + lag2, 
data = toydata, 
offset = logPop, niter = 5000, burn_in = 2500,
rand.int = FALSE, 
covariate.AD = c("lag1", "lag2"),
countsonly = TRUE,
ID.spacetime = c("week", "state"))

### obtain posterior mean/sd and credible intervals 
### for observed counts across week 20-40 
### at states CA and NY 
sum.counts(counts.post = fit$pred.counts,
ID.aggre = "state", timeperiod = 20:40, locs = c("CA", "NY"),
ID.time = "week", ID.loc = "state")

}
