% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Simulated data for the illustration of using spatial-temporal negative-binomial 
regression models to estimate attributable deaths}
\format{
A data frame with 4743 rows and 6 variables:
\describe{
  \item{logPop}{log of populations in each state}
  \item{week}{index of epidemic weeks}
  \item{state}{index of states}
  \item{lag1}{simulated percent positive at lag 1}
  \item{lag2}{simulated percent positve at lag 2}
  \item{y}{observed death counts}
}
}
\usage{
data(toydata)
}
\description{
A dataset containing the simulated death counts and the percent positive across 
51 US states and 93 epidemic weeks. The variables are as follows:
}
\examples{
data(toydata)
}
\keyword{datasets}
