% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmatrix_USstates.R
\docType{data}
\name{W.s}
\alias{W.s}
\title{A first-order adjacency matrix for 50 states and DC in the US}
\format{
A 51 by 51 symmetric matrix named as W.s
}
\usage{
data(adjmatrix_USstates)
}
\description{
A 51 by 51 symmetric matrix named as W.s
}
\examples{
data(adjmatrix_USstates)
}
\keyword{datasets}
